backend_bot/requirements.txt
flask 
flask-socketio
python-dotenv
pydantic
motor
redis
eventlet